Address review comments: move kart_state functions...
into working_copy.py where they are common to both tabular
and file-system working copies.
